Use floats where possible

+++ src/Astap.Lib/Astrometry/CelestialObject.cs
@@ -6,6 +6,6 @@ public readonly record struct CelestialObject(
     double RA, // RA in degrees, 0..24
     double Dec, // Dec in degrees, -90..90 (+ is north)
     Constellation Constellation,
-    double V_Mag, // Visual magnitude or V-mag (in UVB), NaN if not defined
-    double SurfaceBrightness // Surface brightness for galaxies in mag/arcsec^2, NaN if not defined
+    float V_Mag, // Visual magnitude or V-mag (in UVB), NaN if not defined
+    float SurfaceBrightness // Surface brightness for galaxies in mag/arcsec^2, NaN if not defined
 );
